chore: refactor formatting and rendering out of frontend
